fix(glyphs): use same int size for validation
resolves #4454

+++ src/engine/config.go
@@ -316,7 +316,7 @@ func escapeGlyphs(s string, migrate bool) string {
 		}
 
 		if migrate {
-			if val, OK := cp[int(r)]; OK {
+			if val, OK := cp[uint64(r)]; OK {
 				r = rune(val)
 			}
 		}
